requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
PyMuPDF
faiss-cpu
python-dotenv
